POC Export list of assets in CSS based on `url()`
The purpose of this POC is to list the `url` used in CSS values.
This list is exported in Javascript and can be used later to generate
[content preloading header (`<link rel="preload">`)](https://developer.mozilla.org/en-US/docs/Web/HTML/Preloading_content)
or to do some HTTP/2 pushes.

For instance, if your CSS include a background image:

```css
div {
  background: url(/media/background.jpg);
}
```

The following `const` will be exported:

```javascript
export const assets={
  '/media/background.jpg': 'image'
};
```

The object key is the url and the value is the `as` value you should
use. This value is build from CSS keywords, properties, parent...
If not any `as` value is found, it returns `null`.